[ui] Break apart Asset Check page
