+++ js_modules/dagster-ui/packages/ui-core/src/assets/asset-checks/AssetChecksTabs.tsx
@@ -1,0 +1,17 @@
+import {Tab, Tabs} from '@dagster-io/ui-components';
+
+export type AssetChecksTabType = 'overview' | 'execution-history';
+
+interface Props {
+  activeTab: AssetChecksTabType;
+  onChange: (tabId: AssetChecksTabType) => void;
+}
+
+export const AssetChecksTabs = ({activeTab, onChange}: Props) => {
+  return (
+    <Tabs selectedTabId={activeTab} onChange={onChange}>
+      <Tab id="overview" title="Overview" />
+      <Tab id="execution-history" title="Execution history" />
+    </Tabs>
+  );
+};
